requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-mysql==4.2.0
djangorestframework==3.12.4
docopt==0.6.2
idna==3.3
mysql==0.0.3
mysqlclient==2.0.3
pipreqs==0.4.10
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
yarg==0.1.9
gunicorn==20.1.0